Move `LocatedTree` into `tree` submodule
